requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
pillow>=10.0.0
tqdm>=4.60.0
numpy>=1.24.0
